requirements.txt
absl-py==0.7.1
alabaster==0.7.10
alipay-sdk-python==3.3.92
aliyun-python-sdk-core-v3==2.8.9
aliyun-python-sdk-kms==2.5.0
aliyun-python-sdk-sts==3.0.0
altgraph==0.16.1
amqp==2.4.2
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
appdirs==1.4.3
APScheduler==3.6.3
asn1crypto==0.24.0
astor==0.8.0
astroid==2.0.4
astropy==2.0.3
attrs==17.4.0
Automat==0.7.0
autopep8==1.4.3
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
billiard==3.5.0.5
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
bottle==0.12.17
Bottleneck==1.2.1
captcha==0.2.4
celery==4.2.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
constantly==15.1.0
contextlib2==0.5.5
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.3
crcmod==1.7
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
diff-match-patch==20121119
distributed==1.20.2
Django==2.1.7
django-appconf==1.0.3
django-autoslug==1.9.4
django-compressor==2.3
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-debug-toolbar==2.0
django-filter==1.1.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-ipware==2.1.0
django-mdeditor==0.1.14
django-mssql==1.8
django-pglocks==1.0.4
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-pytds==1.5
django-qrcode==0.3
django-rest-swagger==2.2.0
django-reversion==3.0.4
django-silk==3.0.1
django-sqlserver==1.11
django-test-without-migrations==0.6
django-uuslug==1.1.8
django-wechat==0.1a1
djangorestframework==3.8.2
docopt==0.6.2
docutils==0.14
elasticsearch==6.3.1
elasticsearch-dsl==7.0.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Faker==0.8.15
fastcache==1.0.2
filelock==2.0.13
Flask==0.12.2
Flask-Cors==3.0.3
flower==0.9.2
flup==1.0.3
future==0.17.1
gast==0.2.2
geographiclib==1.49
geopy==1.15.0
gevent==1.2.2
glob2==0.6
gprof2dot==2016.10.13
greenlet==0.4.12
grpcio==1.14.1
h5py==2.7.1
heapdict==1.0.0
html2text==2018.1.9
html5lib==1.0.1
httpie==1.0.2
httplib2==0.9.2
hyperlink==18.0.0
idna==2.8
image==1.5.27
imageio==2.2.0
imagesize==0.7.1
incremental==17.5.0
ipaddress==1.0.22
ipykernel==5.1.4
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
iso8601==0.1.12
isort==4.3.4
itsdangerous==0.24
itypes==1.1.0
jdcal==1.4
jedi==0.11.1
jieba==0.39
Jinja2==2.10
jpush==3.3.3
jsonpickle==1.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.4
jupyterlab-launcher==0.10.2
kafka==1.3.5
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Kivy==1.10.1
Kivy-Garden==0.1.4
kivy.deps.glew==0.1.10
kivy.deps.sdl2==0.1.18
kombu==4.4.0
lazy-object-proxy==1.3.1
libsass==0.19.2
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
macholib==1.11
Mako==1.0.13
Markdown==3.1.1
markdown2==2.3.8
MarkupSafe==1.0
matplotlib==2.1.2
mbstrdecoder==0.5.0
mccabe==0.6.1
menuinst==1.4.11
mistune==0.8.4
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
mysqlclient==1.3.12
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
netaddr==0.7.19
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
ntplib==0.3.3
num2words==0.5.10
numba==0.36.2
numexpr==2.6.4
numpy==1.17.4
numpydoc==0.7.0
odfpy==1.3.6
odo==0.5.1
ofxparse==0.20
olefile==0.46
openapi-codec==1.3.2
opencv-python==4.1.1.26
openpyxl==2.5.4
optionaldict==0.1.1
oss2==2.5.0
packaging==19.0
pandas==0.25.3
pandocfilters==1.4.2
parsel==1.5.1
parso==0.1.1
partd==0.3.8
passlib==1.7.1
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pdfkit==0.6.1
pefile==2018.8.8
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prettytable==0.7.2
prompt-toolkit==1.0.15
protobuf==3.8.0
psutil==5.4.3
psycopg2==2.8.3
py==1.5.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.6.1
pycryptodomex==3.7.2
pycurl==7.43.0.1
PyDispatcher==2.0.5
pydot==1.4.1
pyflakes==1.6.0
Pygments==2.4.2
PyHamcrest==1.9.0
PyInstaller==3.4
PyJWT==1.6.4
pylint==2.3.1
pymongo==3.7.2
PyMySQL==0.9.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.4.0
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.4
PySocks==1.6.7
pytesseract==0.2.5
pytest==3.3.2
pytest-runner==4.2
python-dateutil==2.8.0
python-logstash==0.4.6
python-memcached==1.59
python-slugify==3.0.2
python-tds==1.9.1
pytypeutil==0.0.1
pytz==2019.1
pyusb==1.0.2
PyWavelets==0.5.2
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5
PyYAML==3.12
pyzmq==16.0.3
qrcode==6.1
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
queuelib==1.5.0
raven==6.9.0
rcssmin==1.0.6
redis==2.10.6
reportlab==3.5.23
requests==2.14.2
rjsmin==1.1.0
rope==0.10.7
rsa==4.0
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
Scrapy==1.5.1
seaborn==0.8.1
selenium==3.141.0
Send2Trash==1.4.2
serpy==0.3.1
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.15.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
socketIO-client==0.7.2
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
sqlparse==0.3.0
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tablib==0.12.1
tblib==1.3.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
Twisted==18.7.0
typed-ast==1.1.0
typing==3.6.2
tzlocal==2.0.0
unicodecsv==0.14.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.25.3
vine==1.2.0
virtualenv==16.7.6
vobject==0.9.6.1
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.55.0
wechatpy==1.7.0
Werkzeug==0.14.1
WeRoBot==1.9.0
Whoosh==2.7.4
widgetsnbextension==3.1.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wkhtmltopdf==0.2
wrapcache==1.0.8
wrapt==1.10.11
WTForms==2.2.1
xadmin==2.0.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.25.0
zict==0.1.3
zope.interface==4.6.0
